+++ CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.20.0]
+## [0.20.0] - 2018-05-24
 
 ### Added
 
 - Fonts are now part of the html-sketchapp export
 - Copy page files in subfolders
+- Demo content (pages and components) can be generated on initialization
+  - Use the `uiengine init --demo` command
+  - This should give a basic overview for some of the features
+  - Beware: It uses the html adapter, hence the components are very simplistic!
 
 ### Changed